app/requirements.txt
EdgeGPT
python-decouple
falcon
